# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie App API",
        "description": "Tài liệu API cho hệ thống đặt vé xem phim",
        "contact": {
            "email": "youremail@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/phim/{ma_phim}": {
            "get": {
                "tags": [
                    "Phim"
                ],
                "summary": "Lấy thông tin phim theo mã",
                "description": "API này dùng để lấy thông tin chi tiết của một phim dựa trên mã phim.",
                "operationId": "7299464f928292349f2dc51eab28c1c2",
                "parameters": [
                    {
                        "name": "ma_phim",
                        "in": "path",
                        "description": "Mã phim cần tìm",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin phim chi tiết",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ma_phim": {
                                            "type": "string",
                                            "example": "PH001"
                                        },
                                        "ten_phim": {
                                            "type": "string",
                                            "example": "Phim 1"
                                        },
                                        "mo_ta": {
                                            "type": "string",
                                            "example": "Mô tả phim 1"
                                        },
                                        "ngay_cong_chieu": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-04-12"
                                        },
                                        "the_loai": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Hành động"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Phim không tồn tại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phim không tồn tại"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}